requirements.txt
requests
google-cloud-speech
google-cloud-language